Add Unicode Block Elements
The block element characters in Monaco are all sorts of screwed up.
In particular the left half block is narrower than the left
three-eights block and the full block is narrower than the 7/8s block.
The characters are used to display progress bars, which look wonky if
the characters don't do what they say they do. The ones in Menlo, are
fine, so use those here.